Remove Vite and migrate to Vue CLI
