+++ package.json
@@ -3,10 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
-    "test": "vitest run"
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
+    "test": "vue-cli-service test:unit"
   },
   "dependencies": {
     "@fullcalendar/core": "^5.10.1",
@@ -28,16 +27,20 @@
     "xml-js": "^1.6.11"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^4.0.0",
+    "@vue/cli-plugin-babel": "^5.0.0",
+    "@vue/cli-plugin-typescript": "^5.0.0",
+    "@vue/cli-plugin-unit-jest": "^5.0.0",
+    "@vue/cli-service": "^5.0.0",
     "@vue/test-utils": "^2.3.0",
     "eslint": "^7.32.0",
     "eslint-plugin-vue": "^8.0.3",
     "gh-pages": "^3.2.3",
     "jsdom": "^26.1.0",
     "typescript": "^5.0.0",
-    "vite": "^4.0.0",
-    "vitest": "^0.34.0",
-    "vue-tsc": "^1.4.2"
+    "jest": "^27.1.0",
+    "ts-jest": "^27.0.4",
+    "babel-jest": "^27.1.0",
+    "@vue/vue3-jest": "^27.0.0-alpha.3"
   },
   "eslintConfig": {
     "root": true,
