+++ public/index.html
@@ -18,7 +18,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
     <style>
       body{margin:0}
       #app{display:flex;flex-direction:column;min-height:100vh;}
