+++ tsconfig.json
@@ -14,8 +14,7 @@
     "paths": {
       "@/*": ["src/*"]
     },
-    "types": ["vite/client"]
   },
-  "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.vue", "test/**/*.ts"],
+  "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.vue", "test/**/*"],
   "exclude": ["node_modules"]
 }
